Rename detectProjectType util
